+++ src/test/kotlin/com/pestphp/pest/types/ThisFieldTypeTest.kt
@@ -13,8 +13,20 @@ class ThisFieldTypeTest : BaseTypeTest() {
         assertCompletion("a", "b")
     }
 
+    fun testBeforeEachNamespace() {
+        myFixture.configureByFile("tests/beforeEachNamespace.php")
+
+        assertCompletion("a", "b")
+    }
+
     fun testAfterEach() {
         myFixture.configureByFile("tests/afterEach.php")
+
+        assertCompletion("a", "b")
+    }
+
+    fun testAfterEachNamespace() {
+        myFixture.configureByFile("tests/afterEachNamespace.php")
 
         assertCompletion("a", "b")
     }