+++ src/test/resources/com/pestphp/pest/types/thisField/beforeEachNamespace.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\Works;
+
+class Foo {
+    public int $a;
+
+    public function b(){}
+}
+
+beforeEach(function () {
+    $this->foo = new Foo();
+});
+
+it('has home', function () {
+    $this->foo-><caret>
+});