+++ src/test/resources/com/pestphp/pest/types/thisField/beforeEachNamespaceCompletion.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Works;
+
+beforeEach(function () {
+    $this->foo = 1;
+});
+
+it('has home', function () {
+    $this-><caret>
+});